Move build-scripts/relabel-ids.py to ssg/_build_renumber.py
